requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
paramiko==3.5.0
httpx==0.28.0
aiofiles==24.1.0
python-multipart==0.0.20
pytesseract==0.3.13
Pillow==11.1.0
pytest==8.3.4
